move lib files into src subfolder
