+++ backend/src/document-utils/realtimeDocumentUpdates.ts
@@ -1,8 +1,8 @@
 // TO DO: auth checks after first access
 
-import { Document } from "@lib/documentTypes";
-import { OnlineEntity, UpdateType } from "@lib/realtimeUserTypes";
-import { UserEntity } from "@lib/userTypes";
+import { Document } from "@lib/src/documentTypes";
+import { OnlineEntity, UpdateType } from "@lib/src/realtimeUserTypes";
+import { UserEntity } from "@lib/src/userTypes";
 
 import FirebaseWrapper from "../firebase-utils/FirebaseWrapper";
 import { subscribeUserToUserDocumentPool } from "./realtimeOnlineUsers";
