+++ backend/src/document-utils/realtimeOnlineUsers.ts
@@ -1,5 +1,5 @@
-import { OnlineEntity, UpdateType } from "@lib/realtimeUserTypes";
-import { UserEntity } from "@lib/userTypes";
+import { OnlineEntity, UpdateType } from "@lib/src/realtimeUserTypes";
+import { UserEntity } from "@lib/src/userTypes";
 
 import FirebaseWrapper from "../firebase-utils/FirebaseWrapper";
