+++ lib/src/documentTypes.ts
